About Page Reworded
About page has been slightly shortened, please feel free to look over my changes and comment on it. Note that content may be changed when it comes to implementing in website. I have not shortened any of the other pages yet, I have messaged all of you on slack to discuss what will be happening in regards to those.